feat: support referencing csv file in fga store file
